+++ optimum/habana/transformers/models/llama/modeling_llama.py
@@ -1507,7 +1507,7 @@ class GaudiLlamaForCausalLM(LlamaForCausalLM):
 
         # Only compute necessary logits, and do not upcast them to float if we are not computing the loss
         slice_indices = slice(-logits_to_keep, None) if isinstance(logits_to_keep, int) else logits_to_keep
-        logits = self.lm_head(hidden_states[:, slice_indices, :])
+        logits = self.lm_head(hidden_states[:, slice_indices, :]).float()
 
         loss = None
         if labels is not None:
